+++ CHANGELOG.md
@@ -35,6 +35,7 @@
     01. :gift: Added class `dev.orne.config.AbstractConfigCryptoEngine`
     01. :gift: Added class `dev.orne.config.ConfigCryptoAesGcmEngine`
 01. :gift: Added class `dev.orne.config.DefaultConfigCryptoProvider`
+01. :gift: Added class `dev.orne.config.PooledConfigCryptoProvider`
 
 # :package: 0.1.0